Split seed.js into smaller files
